+++ lib/everything_search_engine.dart
@@ -1,3 +1,5 @@
+// GENERATED CODE - DO NOT MODIFY BY HAND
+
 // Copyright (c) 2022 Krysl
 
 /// Automatically generate index / everything_search_engine / library files
@@ -5,8 +7,14 @@
 library everything_search_engine;
 
 export 'src/assets.dart';
+export 'src/ffi/error.dart';
 export 'src/ffi/everything.dart';
+export 'src/ffi/file_attribute.dart';
+export 'src/ffi/file_attribute_constants.dart';
+export 'src/ffi/request_flags.dart';
+export 'src/ffi/sort.dart';
 export 'src/ffi/string.dart';
+export 'src/ffi/target_machine.dart';
 export 'src/ffi/types.dart';
 export 'src/query/query.dart';
 export 'src/query/results.dart';