+++ CHANGELOG.md
@@ -1,3 +1,16 @@
+## 0.1.0
+* add some missing class
+* fix some api's return type to make it more usable
+* make `runQuery` function more powerful
+
+## 0.0.3
+
+* Update doc
+
+## 0.0.2
+
+* Update doc
+
 ## 0.0.1
 
 * Basic query support